requirementsOld.txt
cloudant==2.12.0
Django==3.0.3
django-chartjs==2.0.0
djangorestframework==3.11.0
findspark==1.3.0
gunicorn==20.0.4
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.3.0
ipython==7.9.0
jsonpickle==1.3
jupyter==1.0.0
livereload==2.6.1
pandas==0.25.3
pyspark==2.4.5
python-twitter==3.5
searchtweets==1.7.4
whitenoise==5.0.1
python-dateutil==2.8.1
ibmcloudenv>=0